backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & Web Scraping
httpx==0.24.1
beautifulsoup4==4.12.2
lxml==4.9.3

# AI & LLM
google-generativeai==0.3.2
openai==1.3.7

# Database & Storage
supabase==1.0.4
qdrant-client==1.7.1

# Rate Limiting & Middleware
slowapi==0.1.9
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.24.1  # For testing HTTP clients

# Development
black==23.11.0
isort==5.12.0
